refactor(graph): Stage C.5 — split core/graph_editor.py 20KB → package, all files ≤11KB
Splits the boundary-case 20.1 KB monolithic core/graph_editor.py into
a 3-module package along the helpers / writes / facade seam.
Resolves PLATFORM_READINESS Gate v0.3 dimension A for this file
(5th of five violations after C.1/C.2/C.3/C.4).

Module-size violations: 5 → 0. **Gate A fully cleared.**

## What changed

- core/graph_editor.py (20.1 KB, 497 lines, 11 module-level functions
  + 1 regex constant) → core/graph_editor/ package preserving the
  external surface byte-identical.
- No function signatures, return shapes, or behaviour changed.
  Pure refactor.

## File sizes (post-split)

  __init__.py     4.3 KB   facade + read_relation + graph_edit_enabled
  _helpers.py     6.8 KB   _FM_SPLIT_RE + frontmatter I/O +
                           ontology helpers + relation lookup +
                           _validate_sources
  _writes.py     11.0 KB   replace_relation_sources / append_relation_source
                           / delete_relation (3 bilateral mutation ops)

All three files comfortably under the 20 KB gate.

## Module layout

  L0 helpers (yaml + ontology lazy) → _helpers.py
  L1 writes (use _helpers)         → _writes.py
  L1 facade (re-exports + flag)    → __init__.py

No circular deps. The lazy `core.ontology` import is preserved
verbatim in _helpers (_inverse_type / _label_for_type).

## External surface preserved

All current import paths work byte-identical:

- from core.graph_editor import replace_relation_sources    (server_llmwiki.py)
- from core.graph_editor import append_relation_source       (server_llmwiki.py)
- from core.graph_editor import delete_relation              (server_llmwiki.py)
- from core.graph_editor import read_relation                (server_llmwiki.py)
- from core.graph_editor import graph_edit_enabled           (server_llmwiki.py)
- from core.graph_editor import _load_entity_by_id, _write_entity
                                                             (core/graph_node_editor.py)
- from core.graph_editor import (...)                        (tests/test_phase_e_graph_editor.py)

The package __init__.py re-exports every public + private name that
graph_node_editor + tests pin via direct underscore-import.

## Verification

- tests/test_phase_e_graph_editor.py + test_event_admin_endpoint.py
  → **57/57 pass**
- Full pytest suite (CI ignore list): **1590 passed + 823 subtests
  passed**. Zero regression.
- scripts/dogfood_check.py — 4/4 plugin invariants still PASS.
- ruff check core/graph_editor/ — clean.

## Stage C progress

- C.1 (wiki_generator.py 51 KB) — ✅ PR #427
- C.2 (cascade.py 24 KB) — ✅ PR #428
- C.3 (character_profile.py 24 KB) — ✅ PR #429
- C.4 (security_layer.py 23 KB) — ✅ PR #431
- C.5 (graph_editor.py 20 KB) — ✅ this PR

**All 5 module-size violations cleared. Gate A (architecture < 20 KB)
of PLATFORM_READINESS v0.3 is fully satisfied.**

v0.3 → v0.4 gate remaining: Stage B (CR-E) — the last "Done-when"
criterion. Plus operator-side live re-verification (Stage E.1).

## Out of scope

- core/graph_node_editor.py — separate file, not in the violation list.
- Behavioural changes to edge mutation ops — pure refactor.
